+++ apps/dine-pleiepenger/e2e/playwright/tests/forside.test.ts
@@ -7,7 +7,7 @@ test.beforeEach(async ({ page }) => {
 
 test('Test innhold, funksjonalitet og wcag', async ({ page }) => {
     await page.goto('http://localhost:8080/innsyn');
-    await expect(page.getByRole('heading', { level: 1, name: 'Din pleiepengesak' })).toBeVisible();
+    await expect(page.getByRole('heading', { level: 1, name: 'Din pleiepengesak for sykt barn' })).toBeVisible();
 
     // const søknader = await page.locator('[data-testid="søknadsliste"]').locator('section');
     // expect(await søknader.count()).toBe(3);