+++ apps/dine-pleiepenger/next.config.js
@@ -1,6 +1,7 @@
-const withBundleAnalyzer = require('@next/bundle-analyzer')({
-    enabled: process.env.ANALYZE === 'true',
-});
+// const withBundleAnalyzer = require('@next/bundle-analyzer')({
+//     enabled: process.env.ANALYZE === 'true',
+// });
+
 const { buildCspHeader } = require('@navikt/nav-dekoratoren-moduler/ssr');
 
 const appDirectives = {
@@ -62,4 +63,4 @@ const nextConfig = {
     productionBrowserSourceMaps: true,
 };
 
-module.exports = withBundleAnalyzer(nextConfig);
+module.exports = nextConfig;