+++ apps/dine-pleiepenger/src/utils/__tests__/sakUtils.test.ts
@@ -1,10 +1,14 @@
 import { ISODateToDate } from '@navikt/sif-common-utils';
 import { Behandling } from '../../server/api-models/BehandlingSchema';
 import { Behandlingsstatus } from '../../server/api-models/Behandlingsstatus';
-import { getSisteBehandlingISak, sortBehandlingerNyesteFørst, sortSakshendelse } from '../sakUtils';
-import { Sak } from '../../server/api-models/SakSchema';
+import { harBehandlingSøknadEllerEndringsmelding, sortBehandlingerNyesteFørst, sortSakshendelse } from '../sakUtils';
 import { Sakshendelse, Sakshendelser } from '../../types/Sakshendelse';
 import { Innsendelsestype } from '../../server/api-models/Innsendelsestype';
+import {
+    PleiepengerEndringsmelding,
+    PleiepengerEttersendelse,
+    Pleiepengesøknad,
+} from '../../server/api-models/InnsendelseSchema';
 
 const behandling1: Behandling = {
     status: Behandlingsstatus.UNDER_BEHANDLING,
@@ -28,22 +32,21 @@ const behandling3: Behandling = {
     avsluttetTidspunkt: null,
 };
 
+const innsendtSøknad: Pleiepengesøknad = {
+    innsendelsestype: Innsendelsestype.SØKNAD,
+} as any;
+const innsendtEndringsmelding: PleiepengerEndringsmelding = {
+    innsendelsestype: Innsendelsestype.ENDRINGSMELDING,
+} as any;
+const innsendtEttersendelse: PleiepengerEttersendelse = {
+    innsendelsestype: Innsendelsestype.ETTERSENDELSE,
+} as any;
+
 describe('sakUtils', () => {
     describe('sortBehandlinger', () => {
         it('sorterer behandlinger på sak riktig', () => {
             const sorterteBehandlinger = sortBehandlingerNyesteFørst([behandling1, behandling2, behandling3]);
             expect(sorterteBehandlinger).toEqual([behandling2, behandling3, behandling1]);
-        });
-    });
-
-    describe('getSisteBehandlingISak', () => {
-        it('returnerer siste behandling i sak', () => {
-            const sak: Sak = {
-                saksnummer: '123',
-                behandlinger: [behandling1, behandling2, behandling3],
-            };
-            const sisteBehandling = getSisteBehandlingISak(sak);
-            expect(sisteBehandling).toEqual(behandling2);
         });
     });
 
@@ -75,6 +78,28 @@ describe('sakUtils', () => {
         it('sorterer alltid FORVENTET_SVAR sist', () => {
             const result = [hendelseForventetSvar, hendelse1, hendelse2, hendelse3].sort(sortSakshendelse);
             expect(result).toEqual([hendelse1, hendelse3, hendelse2, hendelseForventetSvar]);
+        });
+    });
+    describe('harBehandlingSøknadEllerEndringsmelding', () => {
+        it('returnerer true hvis behandling har kun søknad', () => {
+            const medSøknad: Behandling = { ...behandling1, innsendelser: [innsendtSøknad] };
+            expect(harBehandlingSøknadEllerEndringsmelding(medSøknad)).toBeTruthy();
+        });
+        it('returnerer true hvis behandling har kun endringsmelding', () => {
+            const medSøknad: Behandling = { ...behandling1, innsendelser: [innsendtEndringsmelding] };
+            expect(harBehandlingSøknadEllerEndringsmelding(medSøknad)).toBeTruthy();
+        });
+        it('returnerer false hvis behandling har kun ettersendelse', () => {
+            const medSøknad: Behandling = { ...behandling1, innsendelser: [innsendtEttersendelse] };
+            expect(harBehandlingSøknadEllerEndringsmelding(medSøknad)).toBeFalsy();
+        });
+        it('returnerer true hvis behandling har søknad + ettersendelse', () => {
+            const medSøknad: Behandling = { ...behandling1, innsendelser: [innsendtSøknad, innsendtEttersendelse] };
+            expect(harBehandlingSøknadEllerEndringsmelding(medSøknad)).toBeTruthy();
+        });
+        it('returnerer true hvis behandling har endringsmelding + ettersendelse', () => {
+            const medSøknad: Behandling = { ...behandling1, innsendelser: [innsendtSøknad, innsendtEndringsmelding] };
+            expect(harBehandlingSøknadEllerEndringsmelding(medSøknad)).toBeTruthy();
         });
     });
 });