+++ apps/omsorgspengesoknad/src/app/søknad/steps/oppsummering/oppsummeringMessages.ts
@@ -1,11 +1,11 @@
 const nb = {
     'steg.oppsummering.søker.header': 'Om deg',
-    'steg.oppsummering.søker.fnr': 'Fødselsnummer: {fødselsnummer}',
+    'steg.oppsummering.søker.fnr': 'Fødselsnummer',
     'steg.oppsummering.barnet.header': 'Om barnet',
-    'steg.oppsummering.barnet.navn': 'Navn: {navn}',
-    'steg.oppsummering.barnet.fødselsdato': 'Fødselsdato: {dato}',
-    'steg.oppsummering.barnet.fnr': 'Fødselsnummer: {fnr}',
-    'steg.oppsummering.barnet.søkersRelasjonTilBarnet': 'Din relasjon til barnet: {relasjon}',
+    'steg.oppsummering.barnet.navn': 'Navn',
+    'steg.oppsummering.barnet.fødselsdato': 'Fødselsdato',
+    'steg.oppsummering.barnet.fnr': 'Fødselsnummer',
+    'steg.oppsummering.barnet.søkersRelasjonTilBarnet': 'Din relasjon til barnet',
     'steg.oppsummering.barnet.sammeAdresse.header': 'Bor du sammen med barnet?',
     'steg.oppsummering.barnet.sammeAdresse.ja': 'Ja',
     'steg.oppsummering.barnet.sammeAdresse.nei': 'Nei',
@@ -25,29 +25,7 @@ const nb = {
 };
 
 const nn: Record<keyof typeof nb, string> = {
-    'steg.oppsummering.søker.header': 'Om deg',
-    'steg.oppsummering.søker.fnr': 'Fødselsnummer: {fødselsnummer}',
-    'steg.oppsummering.barnet.header': 'Om barnet',
-    'steg.oppsummering.barnet.navn': 'Namn: {navn}',
-    'steg.oppsummering.barnet.fødselsdato': 'Fødselsdato: {dato}',
-    'steg.oppsummering.barnet.fnr': 'Fødselsnummer: {fnr}',
-    'steg.oppsummering.barnet.søkersRelasjonTilBarnet': 'Din relasjon til barnet: {relasjon}',
-    'steg.oppsummering.barnet.sammeAdresse.header': 'Bur du saman med barnet?',
-    'steg.oppsummering.barnet.sammeAdresse.ja': 'Ja',
-    'steg.oppsummering.barnet.sammeAdresse.nei': 'Nei',
-    'steg.oppsummering.barnet.sammeAdresse.jaDeltBosted': 'Ja, barnet har delt fast bostad',
-    'steg.oppsummering.barnet.kroniskEllerFunksjonshemmende.header':
-        'Har barnet kronisk/langvarig sjukdom eller funksjonshemming?',
-    'steg.oppsummering.barnet.høyereRisikoForFravær.header':
-        'Har du høgare risiko for fråvær på jobb på grunn av barnet sin sjukdom eller funksjonshemming?',
-    'steg.oppsummering.barnet.høyereRisikoForFraværBeskrivelse.header':
-        'Skildring på korleis barnet sin sjukdom eller funksjonshemming gjev høgare risiko for fråvær frå jobb',
-    'steg.oppsummering.vedlegg.header': 'Vedlegg',
-    'steg.oppsummering.legeerklæring.header': 'Legeerklæring',
-    'steg.oppsummering.samværsavtale.header': 'Avtale om delt fast bostad',
-    'steg.oppsummering.bekrefterOpplysninger':
-        'Eg stadfestar at opplysningane eg har gitt er rette, og at eg ikkje har halde tilbake opplysningar som har tyding for min rett til omsorgspengar.',
-    'validation.harBekreftetOpplysninger.notChecked': 'Du må stadfeste opplysningane.',
+    ...nb,
 };
 export const oppsummeringMessages = {
     nb,