+++ packages/appstatus-react-ds/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @navikt/appstatus-react-ds
 
+## 2.18.11
+
+### Patch Changes
+
+-   a10b53c: Pakkeoppdateringer - minor/patch
+
+## 2.18.10
+
+### Patch Changes
+
+-   b4e0847: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [b4e0847]
+    -   @navikt/sif-common-hooks@0.1.12
+
 ## 2.18.9
 
 ### Patch Changes