+++ packages/sif-common-amplitude/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-amplitude
 
+## 2.25.8
+
+### Patch Changes
+
+-   a10b53c: Pakkeoppdateringer - minor/patch
+
+## 2.25.7
+
+### Patch Changes
+
+-   b4e0847: Pakkeoppdateringer - minor/patch
+
 ## 2.25.6
 
 ### Patch Changes