+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-forms-ds
 
+## 8.23.4
+
+### Patch Changes
+
+-   a10b53c: Pakkeoppdateringer - minor/patch
+
+## 8.23.3
+
+### Patch Changes
+
+-   b4e0847: Pakkeoppdateringer - minor/patch
+
 ## 8.23.2
 
 ### Patch Changes