+++ packages/sif-common-soknad-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-soknad-ds
 
+## 10.15.12
+
+### Patch Changes
+
+-   a10b53c: Pakkeoppdateringer - minor/patch
+
+## 10.15.11
+
+### Patch Changes
+
+-   b4e0847: Pakkeoppdateringer - minor/patch
+
 ## 10.15.10
 
 ### Patch Changes