+++ packages/sif-common-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-utils
 
+## 3.46.6
+
+### Patch Changes
+
+-   a10b53c: Pakkeoppdateringer - minor/patch
+
+## 3.46.5
+
+### Patch Changes
+
+-   b4e0847: Pakkeoppdateringer - minor/patch
+
 ## 3.46.4
 
 ### Patch Changes